Struct slog_term::TermDecorator

pub struct TermDecorator { /* fields omitted */ }

Decorator implemented using term crate

This decorator will add nice formatting to the logs it's outputting. It's based on term crate.

It does not deal with serialization so is !Sync. Run in a separate thread with slog_async::Async.

Methods

impl TermDecorator

pub fn new() -> TermDecoratorBuilder

Start building TermDecorator

pub fn level_to_color(level: Level) -> u16

Level color

Standard level to Unix color conversion used by TermDecorator